Fix annotation handling in signature.py
