+++ src/dags/__init__.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
+from dags.annotations import get_annotations, get_free_arguments
 from dags.dag import (
     concatenate_functions,
     create_dag,
     get_ancestors,
-    get_annotations,
-    get_free_arguments,
 )
 from dags.exceptions import (
     AnnotationMismatchError,
